map-component update, renamed loader-screen
