+++ project/src/components/loader/loader.module.scss
